+++ backend/canisters/community/impl/src/updates/register_poll_vote.rs
@@ -4,7 +4,8 @@ use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use chat_events::{RegisterPollVoteArgs, RegisterPollVoteResult};
 use community_canister::register_poll_vote::{Response::*, *};
-use types::Achievement;
+use types::{Achievement, Chat, TotalVotes};
+use user_canister::{CommunityCanisterEvent, MessageActivity, MessageActivityEvent};
 
 #[update(candid = true, msgpack = true)]
 #[trace]
@@ -62,13 +63,21 @@ fn register_poll_vote_impl(args: Args, state: &mut RuntimeState) -> Response {
     });
 
     match result {
-        RegisterPollVoteResult::Success(votes) => {
+        RegisterPollVoteResult::Success(votes, creator) => {
+            state.data.user_event_sync_queue.push(
+                creator,
+                CommunityCanisterEvent::MessageActivity(MessageActivityEvent {
+                    chat: Chat::Channel(state.env.canister_id().into(), channel.id),
+                    thread_root_message_index: args.thread_root_message_index,
+                    message_index: args.message_index,
+                    activity: MessageActivity::PollVote,
+                    timestamp: now,
+                    user_id: matches!(votes.total, TotalVotes::Visible(_)).then_some(user_id),
+                }),
+            );
+
             if args.new_achievement {
-                state.data.achievements.notify_user(
-                    user_id,
-                    vec![Achievement::VotedOnPoll],
-                    &mut state.data.fire_and_forget_handler,
-                );
+                state.data.notify_user_of_achievement(user_id, Achievement::VotedOnPoll);
             }
 
             handle_activity_notification(state);