+++ backend/integration_tests/src/utils.rs
@@ -3,7 +3,7 @@ use pocket_ic::PocketIc;
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use std::time::SystemTime;
 use std::{path::PathBuf, time::UNIX_EPOCH};
-use types::{Hash, TimestampMillis};
+use types::{Hash, TimestampMillis, TimestampNanos};
 
 pub fn principal_to_username(principal: Principal) -> String {
     principal.to_string()[0..5].to_string()
@@ -19,7 +19,7 @@ pub fn now_millis(env: &PocketIc) -> TimestampMillis {
     now_nanos(env) / 1_000_000
 }
 
-pub fn now_nanos(env: &PocketIc) -> TimestampMillis {
+pub fn now_nanos(env: &PocketIc) -> TimestampNanos {
     env.get_time().duration_since(SystemTime::UNIX_EPOCH).unwrap().as_nanos() as u64
 }