+++ backend/canisters/airdrop_bot/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.1394](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1394-airdrop_bot)] - 2024-10-16
+
 ### Changed
 
 - Serialize large integers as strings when using MessagePack ([#6315](https://github.com/open-chat-labs/open-chat/pull/6315))